+++ src/utils/premadeTemplateGenerators/createReuseableComponent.ts
@@ -1,9 +1,9 @@
-import { createComponentBasicFiles } from '../createComponentBasicFiles';
+import { createComponentWithTypesAndStyle } from '../createComponentFile';
 import { createStoryFile } from '../createStoryFile';
 import { createTestFileAndUtils } from '../createTestFileAndUtils';
 
 export const createReusableComponent = async (destinationPath: string, componentName: string) => {
-  await createComponentBasicFiles(destinationPath, componentName);
+  await createComponentWithTypesAndStyle(destinationPath, componentName);
   await createStoryFile(destinationPath, componentName);
   await createTestFileAndUtils(destinationPath, componentName);
   // TODO