+++ CHANGELOG.md
@@ -7,9 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-## 0.1.1 - 2022-10-22
 ### Added
+
+- Support for custom template including test, types, style and component
+- Added types to Component with test premade template
+
+### Changed
+
+- Updated strings to use single quote
+
+## 0.1.1 - 2022-10-22
+
+### Added
+
 - Pre made templates: `Reuseable component` and `Component with test`
 
 ### Removed
+
 - Support for custom template selection